lab/scripts/k8s/horseKeeper/requirements.txt
kubernetes==26.1.0
PyYAML~=6.0
setuptools~=65.5.1
requests~=2.28.2
beautifulsoup4~=4.12.0